+++ main_test.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/ProRocketeers/yoke-chart/schema"
 	"github.com/lithammer/dedent"
+	"github.com/stretchr/testify/assert"
 )
 
 // meant for testing the parsing mechanism and custom validation logic etc.
@@ -17,7 +18,74 @@ func TestMain(t *testing.T) {
 		Asserts func(*testing.T, schema.InputValues, error)
 	}
 
-	cases := map[string]CaseConfig{}
+	cases := map[string]CaseConfig{
+		"passes empty Kind": {
+			Input: `
+        namespace: foo
+        service: foo
+        component: bar
+        environment: test
+
+        image:
+          repository: foo
+          tag: bleh
+      `,
+			Asserts: func(t *testing.T, iv schema.InputValues, err error) {
+				assert.NoError(t, err)
+			},
+		},
+		"passes Kind = Deployment": {
+			Input: `
+        namespace: foo
+        service: foo
+        component: bar
+        environment: test
+
+        image:
+          repository: foo
+          tag: bleh
+
+        kind: Deployment
+      `,
+			Asserts: func(t *testing.T, iv schema.InputValues, err error) {
+				assert.NoError(t, err)
+			},
+		},
+		"passes Kind = StatefulSet": {
+			Input: `
+        namespace: foo
+        service: foo
+        component: bar
+        environment: test
+
+        image:
+          repository: foo
+          tag: bleh
+
+        kind: StatefulSet
+      `,
+			Asserts: func(t *testing.T, iv schema.InputValues, err error) {
+				assert.NoError(t, err)
+			},
+		},
+		"fails a non-enum Kind value": {
+			Input: `
+        namespace: foo
+        service: foo
+        component: bar
+        environment: test
+
+        image:
+          repository: foo
+          tag: bleh
+
+        kind: foo
+      `,
+			Asserts: func(t *testing.T, iv schema.InputValues, err error) {
+				assert.Error(t, err)
+			},
+		},
+	}
 
 	for testName, tc := range cases {
 		t.Run(testName, func(t *testing.T) {